pyproject.toml
[project]
name = "datathon-polyfinances"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.64",
    "langdetect>=1.0.9",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "trafilatura>=1.6.1",
]
